+++ docs/legacy_identifier_transfer/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this gear are documented in this file.
 
-## 1.2.0
+## Unreleased
 * Updates to use new identifier lambda functions with restructured identifiers database
   
 ## 1.1.6